pyproject.toml
[project]
name = "halloween-prize-game"
version = "0.1.0"
description = "Raspberry Pi 5 powered prize game to hand out halloween candy"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pygame",
    "rpi-lgpio",
]
